application/ctf_club/views.py: Make CodeQL happy, and also fix some possible errors from happening.
CodeQL isn't able to realize that I'm only only allowing filenames that match
the characters of the set [A-Za-z\-_0-9.] in the filename. Thus it's
impossible for anyone to actually do directory traversal on me as the
character isn't even allowed to let see the directory above it but that's it.
Since I don't allow for \ nor / in the filename. So they'd read nothing at worst
they'd be able to cause an IsADirectoryError. That's it. So to make CodeQL happy
I'm patching it. Further I make sure that it is a file and that it exists before
attempting to return it preventing any errors from occuring.